Fix ID mapping issue: change int to Long in SupportTicket model

+++ src/main/java/com/yigitaslan/call_center_system/repository/ISupportTicketRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ISupportTicketRepository extends JpaRepository<SupportTicket, Integer> {
+public interface ISupportTicketRepository extends JpaRepository<SupportTicket, Long> {
 }
